Add files via upload
Process of analyzing and summarizing a dataset in order to gain insights and understand the relationships between variables. EDA typically involves visualizing the data, identifying patterns and trends, and looking for any anomalies or issues in the data that may need to be addressed. The goal of EDA is to provide a better understanding of the data, which can help inform further analysis or modeling. It also includes pre-processing and post-processing.